+++ pkg/passes/invalid-return.pm
@@ -1,0 +1,23 @@
+return 1
+
+def my_function()
+    return None
+end
+
+gen my_generator()
+    return None
+end
+
+for a in range(100)
+    return 2
+end
+
+for value in range(2000)
+    def __anonymous()
+        return value
+    end
+end
+
+if False
+    return 3
+end